requirements.txt
Flask==3.0.0
Flask-Compress==1.17
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
deezer-py
git+https://gitlab.com/RemixDev/deemix-py.git

